Add hnswlib
hnswlib is a header-only library that provides an implementation of
approximate nearest neighbor.

A specific commit from Git is used. The newest release does not compile
on Ubuntu due to a missing #include of assert.h. Once that's merged into
master and into a release, new_git_repository should be swapped with
http_archive.