Switch bazel gRPC library to the primary repo
Prior to this we were using a repo that had specialized rules for
bazel, but it's become a little out of date. It's easier to just
use the main repo directly, but it also makes it slightly more
annoying to use with bazel since it doesn't have the nice wrapper
that we had prior to this.