Add nearest_neighbor lookup by embedding
Prior to this, nearest neighbor lookups were only able to be performed
by key. This meant that you couldn't do a nearest neighbor lookup on
something that wasn't already stored in the embedding store. This
is useful in use cases like a search engine where the NN query may
be an embedding of a search query and the embedding store stores
the documents.