if no records are present, return an empty list
before the refactor, get_metadata_list would use the get_type_table() function which would gracefully return empty results if none are found.

with the new api, it makes use of query_resource() which throws an error if no records are found (a valid scenario). removing the error and instead returning an empty list.